Kyle Nierras Update
