+++ lib/Custom_Widgets/Custom_Dialog.dart
@@ -5,8 +5,10 @@ class CustomDialog extends StatelessWidget {
   final VoidCallback onSignOut;
   final String message;
   final bool? showtitle; 
+  final String buttonText;
 
-  CustomDialog({
+  const CustomDialog({super.key, 
+    required this.buttonText,
     required this.onSignOut,
     required this.message,
     this.showtitle,
@@ -16,16 +18,29 @@ class CustomDialog extends StatelessWidget {
   Widget build(BuildContext context) {
     return CupertinoAlertDialog(
       title: showtitle ?? false ? const Text('Message') : null,
-      content: Text(message),
+      content: Text(
+        message,
+        style: const TextStyle(
+          fontWeight: FontWeight.bold, // Set the content text to bold
+          fontSize: 17,
+        ),
+      ),
       actions: <Widget>[
         CupertinoDialogAction(
+          textStyle: const TextStyle(
+            color: Color.fromARGB(255, 0, 0, 255)
+          ),
           child: const Text("Cancel"),
           onPressed: () {
             Navigator.of(context).pop(); // Close the dialog
           },
         ),
         CupertinoDialogAction(
-          child: const Text('Yes'),
+          textStyle: const TextStyle(
+            fontWeight: FontWeight.bold,
+            color: Color.fromARGB(255, 255, 0, 0),
+          ),
+          child: Text(buttonText),
           onPressed: () {
             FirebaseAuth.instance.signOut();
             onSignOut();
